feat: no modal on company verification
